requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
construct==2.5.3
defusedxml==0.5.0
diff-match-patch==20181111
Django==1.11
django-celery==3.2.2
django-crispy-forms==1.7.2
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.1.0
django-redis==4.10.0
django-reversion==3.0.2
djangorestframework==3.9.0
elasticsearch==2.4.0
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.12.0
jdcal==1.4
kombu==3.0.37
odfpy==1.4.0
openpyxl==2.5.12
pefile==2018.8.8
Pillow==5.3.0
python-dateutil==2.7.5
python-ptrace==0.9.3
pytz==2018.7
PyYAML==3.13
redis==2.10.6
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.24.1
xlrd==1.1.0
xlwt==1.3.0
